% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_breakpoint.R
\name{get_envir}
\alias{get_envir}
\title{Find Environment Using Label and Get It}
\usage{
get_envir(envir_label, current_env)
}
\arguments{
\item{envir_label}{label of environment in which lives object to which breakpoint
will be inserted. Retrieved using 'rlang::env_label()'.}

\item{current_env}{'rlang::current_env()' will be passed as argument to this parameter
to ensure that not 'shinybreakpoint' environment, but user environment will be used.}
}
\value{
Environment having the same label as searched.
}
\description{
Search across parent environments, starting from current environment to find
and retrieve environment of chosen label.
}
\details{
To get environment, function has to be run in user's environment, but to find
searched environment, we need to know what to search and for this label is used.
Environment is necessary to be sure that proper object will be found.
}
