% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_filtering_by_id.R
\name{set_filtering_by_id}
\alias{set_filtering_by_id}
\title{Set Option to Filtering by Id}
\usage{
set_filtering_by_id()
}
\value{
Used for side effect - enable \link{reactlog}
and set up cleaning in temporary directory.
}
\description{
Enable \link{reactlog} functionality and remove
data saved in temporary directory when app stops.
}
\details{
This function must be used outside of the \code{server}
part of the app.
}
\examples{
\dontrun{

library(shiny)

shinybreakpoint::set_filtering_by_id() # TODO: remove

appServer <- function(input, output, session) {
  observe({
    input$num
  }, label = "observe_print_num_input")
}

shinyApp(
  ui = fluidPage(
    theme = bslib::bs_theme(5),
    numericInput("num", "Num", 0)
  ),
  server = function(input, output, session) {
    shinybreakpoint::shinybreakpointServer() # TODO: remove
    appServer(input, output, session)
  }
)
}
}
