% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinybreakpoints_mod.R
\name{shinybreakpointsUI}
\alias{shinybreakpointsUI}
\title{UI Module}
\usage{
shinybreakpointsUI(id)
}
\arguments{
\item{id}{from 'shinybreakpointsServer'. Can be chosen by user.}
}
\value{
HTML script tag with JavaScript code.
}
\description{
Add keydown listener.
}
