% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinybreakpoint_mod.R
\name{shinybreakpointUI}
\alias{shinybreakpointUI}
\title{UI Module}
\usage{
shinybreakpointUI(id)
}
\arguments{
\item{id}{from 'shinybreakpointServer'. Can be chosen by user.}
}
\value{
HTML script tag with JavaScript code - returns
key pressed.
}
\description{
Add keydown listener.
}
