% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinybrowser_mod.R
\name{shinybrowserUI}
\alias{shinybrowserUI}
\title{UI Module}
\usage{
shinybrowserUI(id)
}
\arguments{
\item{id}{from 'shinybrowserServer'. Can be chosen by user.}
}
\value{
HTML script tag with JavaScript code.
}
\description{
Add keydown listener.
}
