% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippet.R
\name{snippet}
\alias{snippet}
\title{Write (Append) Basic Shiny App With Shinybreakpoint Functionality}
\usage{
snippet()
}
\value{
Appends code skeleton to the file.
}
\description{
Append code skeleton with \code{shinybreakpoint} functionality to the
active (opened) file in RStudio.
}
\details{
\code{shinybreakpoint} needs server logic to be included in the function separated
from the 'server' part of app and also was developed having Bootstrap
version 5 in mind. This snippet takes care of that.

Snippet is also available as the addin.
}
\examples{

\dontrun{
# use shinybreakpoint::snippet() in the Console in RStudio
# if there is an opened file (in Source Editor),
# then this skeleton should be added to the file:
library(shiny)

ui <- fluidPage(
  theme = bslib::bs_theme(5),
)

appServer <- function(input, output, session) {

}

server <- function(input, output, session) {
  shinybreakpoint::shinybreakpointServer() # TODO: remove
  appServer(input, output, session)
}

shinyApp(ui, server)
}
}
