% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippet.R
\name{snippet}
\alias{snippet}
\title{Write Shiny Snippet With Shinybreakpoint Functionality}
\usage{
snippet(append = TRUE)
}
\arguments{
\item{append}{should the code be added to the existing content of file or replace it? Added (\code{TRUE}) by default.}
}
\value{
Writes code to the file.
}
\description{
Add code skeleton with \code{shinybreakpoint} functionality to the
active (opened) file in RStudio if this file is saved.
}
\details{
\code{shinybreakpoint} needs server logic to be included in the function separated
from the 'server' part of app and also was developed having latest Bootstrap
version in mind. This snippet takes care of that.
}
\examples{

\dontrun{
# use shinybreakpoint::snippet() in the Console in RStudio
# if there is an opened and saved file (in Source Editor),
# then this skeleton should be added to the file:
library(shiny)

ui <- fluidPage(
  theme = bslib::bs_theme(),
)

appServer <- function(input, output, session) {

}

server <- function(input, output, session) {
  appServer(input, output, session)
  shinybreakpoint::shinybreakpointServer() # TODO: remove
}

shinyApp(ui, server)
}
}
