% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippet.R
\name{snippet}
\alias{snippet}
\title{Write (Append) Basic Shiny App With Shinybreakpoint Functionality}
\usage{
snippet()
}
\value{
Appends code skeleton to the file.
}
\description{
Append code skeleton with \code{shinybreakpoint} functionality to the
active (opened) file in RStudio.
}
\details{
\code{shinybreakpoint} needs server logic to be included in the function separated
from the \code{server} part of app, was also developed having Bootstrap
version 5 in mind and some functionalities needs \link[shiny:reactlog]{shiny::reactlog} enabled.
This snippet takes care of that.

Snippet is also available as the addin, under the name \code{snippet}.
}
\examples{

\dontrun{
# use shinybreakpoint::snippet() in the Console in RStudio
# (or find `snippet` addin in Addins menu)
# if there is an opened file (in Source Editor),
# then this skeleton should be added to the file:
library(shiny)

options(shiny.reactlog = TRUE) # TODO: remove

ui <- fluidPage(
  theme = bslib::bs_theme(5),

)

appServer <- function(input, output, session) {

}

server <- function(input, output, session) {
  shinybreakpoint::shinybreakpointServer() # TODO: remove
  appServer(input, output, session)
}

shinyApp(ui, server)
}
}
